Fix testTimeDelta and database seed value to support bigints

+++ simoc_server/database/db_model.py
@@ -74,7 +74,7 @@ class AgentModelState(BaseEntity):
     grid_width = db.Column(db.Integer, nullable=False)
     grid_height = db.Column(db.Integer, nullable=False)
     model_time = db.Column(db.Interval, nullable=False)
-    seed = db.Column(db.Integer, nullable=False)
+    seed = db.Column(db.BigInteger, nullable=False)
     random_state = db.Column(db.PickleType, nullable=False)
